ROLLBACK: AIML Test Unsuccessful
AIML Parser seems to have difficulties with the parsing of the german aiml-file.
The converting Pearl-script aiml2rs.pl will help to convert into RiveScript language to perform the last test.

AIML would be straight forward and easy to use, but the interpreter doesn't choose the right answer. Implementation with AIML might be too risky!